renaming

+++ NeuralNetRanking/crossvalidation.py
@@ -70,7 +70,7 @@ def predict_folder_content(input_folder,model):
         i+=1
         if len(batch)>=mini_batch_size:
 
-            results_of_batch = model(batch)#[0].
+            results_of_batch = model(torch.DoubleTensor(batch))#[0].
             for j,row in enumerate(results_of_batch):
                 result = row[0].data[0].item()
                 results[int(index[j])] = result
